+++ packages/server/src/Hocuspocus.ts
@@ -25,6 +25,7 @@ import {
   onStoreDocumentPayload,
 } from './types.js'
 import { getParameters } from './util/getParameters.js'
+import { useDebounce } from './util/debounce'
 
 export const defaultConfiguration = {
   name: null,
@@ -67,6 +68,8 @@ export class Hocuspocus {
 
   debugger = new Debugger()
 
+  debounce = useDebounce()
+
   constructor(configuration?: Partial<Configuration>) {
     if (configuration) {
       this.configure(configuration)
@@ -105,6 +108,7 @@ export class Hocuspocus {
       connected: this.configuration.connected,
       onAuthenticate: this.configuration.onAuthenticate,
       onLoadDocument: this.configuration.onLoadDocument,
+      afterLoadDocument: this.configuration.afterLoadDocument,
       beforeHandleMessage: this.configuration.beforeHandleMessage,
       beforeBroadcastStateless: this.configuration.beforeBroadcastStateless,
       onStateless: this.configuration.onStateless,
@@ -113,6 +117,7 @@ export class Hocuspocus {
       afterStoreDocument: this.configuration.afterStoreDocument,
       onAwarenessUpdate: this.configuration.onAwarenessUpdate,
       onRequest: this.configuration.onRequest,
+      afterUnloadDocument: this.configuration.afterUnloadDocument,
       onDisconnect: this.configuration.onDisconnect,
       onDestroy: this.configuration.onDestroy,
     })
@@ -198,9 +203,14 @@ export class Hocuspocus {
       // Only run this if the document has finished loading earlier (i.e. not to persist the empty
       // ydoc if the onLoadDocument hook returned an error)
       if (!document.isLoading) {
-        this.debounce(`onStoreDocument-${document.name}`, () => {
-          this.storeDocumentHooks(document, hookPayload)
-        }, this.configuration.unloadImmediately)
+        this.debounce(
+          `onStoreDocument-${document.name}`,
+          () => {
+            this.storeDocumentHooks(document, hookPayload)
+          },
+          this.configuration.unloadImmediately ? 0 : this.configuration.debounce,
+          this.configuration.maxDebounce,
+        )
       } else {
         // Remove document from memory immediately
         this.unloadDocument(document)
@@ -236,48 +246,19 @@ export class Hocuspocus {
     // If the update was received through other ways than the
     // WebSocket connection, we don’t need to feel responsible for
     // storing the content.
-    if (!connection) {
+    // also ignore changes incoming through redis connection, as this would be a breaking change (#730, #696, #606)
+    if (!connection || (connection as unknown as string) === '__hocuspocus__redis__origin__') {
       return
     }
 
-    this.debounce(`onStoreDocument-${document.name}`, () => {
-      this.storeDocumentHooks(document, hookPayload)
-    })
-  }
-
-  timers: Map<string, {
-    timeout: NodeJS.Timeout,
-    start: number
-  }> = new Map()
-
-  /**
-   * debounce the given function, using the given identifier
-   */
-  debounce(id: string, func: Function, immediately = false) {
-    const old = this.timers.get(id)
-    const start = old?.start || Date.now()
-
-    const run = () => {
-      this.timers.delete(id)
-      func()
-    }
-
-    if (old?.timeout) {
-      clearTimeout(old.timeout)
-    }
-
-    if (immediately) {
-      return run()
-    }
-
-    if (Date.now() - start >= this.configuration.maxDebounce) {
-      return run()
-    }
-
-    this.timers.set(id, {
-      start,
-      timeout: setTimeout(run, this.configuration.debounce),
-    })
+    this.debounce(
+      `onStoreDocument-${document.name}`,
+      () => {
+        this.storeDocumentHooks(document, hookPayload)
+      },
+      this.configuration.debounce,
+      this.configuration.maxDebounce,
+    )
   }
 
   /**
@@ -354,15 +335,10 @@ export class Hocuspocus {
   }
 
   storeDocumentHooks(document: Document, hookPayload: onStoreDocumentPayload) {
-    this.hooks('onStoreDocument', hookPayload)
-      .catch(error => {
-        if (error?.message) {
-          throw error
-        }
-      })
+    return this.hooks('onStoreDocument', hookPayload)
       .then(() => {
         this.hooks('afterStoreDocument', hookPayload).then(() => {
-        // Remove document from memory.
+          // Remove document from memory.
 
           if (document.getConnectionsCount() > 0) {
             return
@@ -370,6 +346,13 @@ export class Hocuspocus {
 
           this.unloadDocument(document)
         })
+      })
+      .catch(error => {
+        console.error('Caught error during storeDocumentHooks', error)
+
+        if (error?.message) {
+          throw error
+        }
       })
   }